requirements.txt
apscheduler
peewee
tushare
openai
jinja2
python-dotenv
requests
pyyaml
akshare
